Organize all AR logic using Provider pattern
